+++ app/Forms/Types/ItemType.php
@@ -43,6 +43,7 @@ class ItemType extends AbstractType
             ->add('related_work_order')
             ->add('related_work_total')
             ->add('acquisition_date')
+            ->add('credit')
             ->add('primary_image', FileType::class, [
                 'mapped' => false,
                 'required' => $options['new'],