+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file[^1].
 ## [Unreleased]
 ### Added
 - Laravel Deployer config
+- Item credit attribute
 
 ### Fixed
 - facebook sharing handling, pinterest media preview